combine CsvWriter into CsvDataWriter
