+++ source/Sylvan.Data.Csv/CsvDataReader.cs
@@ -8,7 +8,6 @@ using System.Diagnostics;
 using System.Globalization;
 using System.IO;
 using System.Runtime.CompilerServices;
-using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 
@@ -109,12 +108,6 @@ namespace Sylvan.Data.Csv
 		readonly BinaryEncoding binaryEncoding;
 		readonly bool hasHeaders;
 		readonly StringFactory stringFactory;
-
-
-
-
-
-
 
 		private CsvDataReader(TextReader reader, CsvDataReaderOptions? options = null)
 		{
