+++ source/Sylvan.Data.Csv/Sylvan.Data.Csv.csproj
@@ -3,7 +3,7 @@
   <PropertyGroup>
     <TargetFrameworks>netstandard2.0;netstandard2.1</TargetFrameworks>
     <VersionPrefix>0.10.0</VersionPrefix>
-    <VersionSuffix>B0008</VersionSuffix>
+    <VersionSuffix>B0009</VersionSuffix>
     <Description>A .NET library for reading and writing delimited CSV data.</Description>
     <PackageTags>csv;delimited;datareader</PackageTags>
     <Nullable>enable</Nullable>
